+++ src/tripweaver/conversation/__init__.py
@@ -1,12 +1,27 @@
 """Stateful, locally-replanned travel planning conversations."""
 
 from .hybrid import HybridConversationPlanningService
-from .models import ConversationEvent, PlanDiff, PlanningSession, RevisionIntent, SessionStatus
+from .models import (
+    ConversationEvent,
+    ExecutionTraceStep,
+    PlanDiff,
+    PlanningSession,
+    PlanningSnapshot,
+    ProviderSnapshot,
+    RevisionIntent,
+    SessionMode,
+    SessionStatus,
+    TraceCategory,
+    TraceStatus,
+    TraceSummary,
+)
 from .parser import DeterministicRevisionParser, UnsafeRevisionError
 from .service import ConversationPlanningService, SessionNotFoundError
 
 __all__ = [
     "ConversationEvent", "ConversationPlanningService", "DeterministicRevisionParser",
-    "HybridConversationPlanningService", "PlanDiff", "PlanningSession", "RevisionIntent", "SessionNotFoundError",
-    "SessionStatus", "UnsafeRevisionError",
+    "ExecutionTraceStep", "HybridConversationPlanningService", "PlanDiff", "PlanningSession",
+    "PlanningSnapshot", "ProviderSnapshot", "RevisionIntent", "SessionMode",
+    "SessionNotFoundError", "SessionStatus", "TraceCategory", "TraceStatus", "TraceSummary",
+    "UnsafeRevisionError",
 ]